feat: Add all stop locations to the map
Also refactors some models into the models module in the shared KMP,
along with renaming some of the nearby stops models to be more general.
I was having trouble getting tests which access the internal map data
to work, so that's not included in this. It's probably possible, and
we'll probably want to figure out how as we continue to iterate on map
display. Unfortunately the Mapbox iOS docs have 0 details on testing.