fix: Load correct nearby location on launch
When the app was started with location permissions already granted,
nearby transit was instead loading the default camera location, rather
that the phone's current GPS location. The NearbyTransitLocationProvider
was implemented in a silly way where it was effectively static, and on
app startup, `currentLocation` and `viewportProvider.viewport` were both
changing and running their `onChange` at the same time, so the current
location was getting overridden by the default camera state.