fix: Handle long prediction text better
In the case of "Stopped X stops away", even if there was enough room to
display it all on one line, since the headsign was greedily taking up
space, it would be broken onto two lines. This makes sure that they both
take up as much space as they need when it's available.