feat: Nearby Transit route card UI styling (#172)
* refactor: Rename "Assets" bundle to "Icons"

* feat: Add color assets

* feat: Add mode and FA icon assets

* feat: UI implementation for nearby transit card

* test: Fix tests based on layout changes

* fix: Fix multiple chevrons and misplaced chevrons

* fix: Don't crash when filter has a missing route

* wip: Frameless

* fix: Switch "Bus" text to be part of string

* Revert "fix: Don't crash when filter has a missing route"

This reverts commit 65824de6cff4842aac92259e4f5f28ff08df7c77.

* fix: Set route name maxHeight to icon height

* fix: Handle missing filter routes

This was regularly causing crashes for me while testing. This probably
isn't the ideal way to handle this, but it stops the crashes for now.

* fix: Prevent mismatched list backgrounds

* fix: Better support for prediction localization

Uses an AttributedString to conditionally highlight the formatting of
any numbers in the string, while leaving the rest as body text.

* fix: Use Dividers instead of Rectangles

* fix: Remove unnecessary ViewBuilder

* fix: Add formatting to Bus string

The design called for "Bus" to not be bold, implementing this with
separate Text views was clunky and caused layout issues, but using
AttributedStrings, we can do this in a single Text!

* test: Revert tests that were relying on min

Now that we're using AttributedStrings here, we need to be testing for
the full string including "min" again.

* fix: Switch to a LazyVStack in the ScrollView

* fix: Handle long prediction text better

In the case of "Stopped X stops away", even if there was enough room to
display it all on one line, since the headsign was greedily taking up
space, it would be broken onto two lines. This makes sure that they both
take up as much space as they need when it's available.