feat: Add separate touch target padding layer
This is so that tapping on a stop icon always brings you to that stop,
even if it overlaps with the padding of a nearby stop. If we put the
padding directly on the stop layer, there are some cases where the
padding of one stop totally covers a nearby stop's icon, which makes it
impossible to select that nearby stop, even if you tap directly on its
icon. Since mapbox needs an icon image to render a tap target, I added
a single transparent pixel SVG to serve as the dummy stop icon.