feat: Update stop icons on the map (#207)
* assets: Add stop map icons

* feat: Dark theme map 🌃

* feat: Route stop colors and transfer stop icons

Recolors stops based on the routes that they serve, with separate icons
at wide and close zoom levels. Also creates extra large transfer stop
icons, which contain separate nested icons for all the routes served at
that stop.

* feat: Update nearby location search indicator

* feat: Update recenter button styling

* feat: Resize stop icons by mode when zooming

Bus stops shrink to remain legible, commuter rail stops increase in size
at very wide zoom levels, and subway/SL increase their size mid-wide.

* feat: Include canonical stops

* feat: Update close and default zoom thresholds

* fix: Hide shuttle stops & update SL display logic

* feat: Add separate touch target padding layer

This is so that tapping on a stop icon always brings you to that stop,
even if it overlaps with the padding of a nearby stop. If we put the
padding directly on the stop layer, there are some cases where the
padding of one stop totally covers a nearby stop's icon, which makes it
impossible to select that nearby stop, even if you tap directly on its
icon. Since mapbox needs an icon image to render a tap target, I added
a single transparent pixel SVG to serve as the dummy stop icon.

* fix: Use different icons for bus at transfer stops

* feat: Add ferry icons

* test: Add tests for GlobalStaticData mapStops

* feat: Add Mattapan trolley map icons

* test: Get existing tests passing

* test: Add test for checking stop route prop

Also rename testDirectionFilter, which xcode was picking up as a test
case because the name started with "test"

* refactor: Move zoom defaults into separate struct

* refactor: Move global map data into separate class

Also fix the issue of map icons disappearing when the app gets
backgrounded.

* fix: Simplify stop symbol sort logic

* revert: Revert unrelated test helper func rename