fix: Move bus map icons onto separate layer
Splits out a separate stop and alert layer just for bus stops. This was
done to prevent bus alert icons from being layered between rail stop
icons and rail alerts. I also needed to remove the opacity transition
for stop layers, because the updates on every camera move were making
the map prohibitively slow. There's probably a smarter way to update it,
but we can add that back later if we need to.