refactor: Hoist stop filter state
It wasn't possible for the map to respond to the stop detail page filter
changing, because that state was only contained deeper within the stop
details page, so the map couldn't access updates. This moves the filter
state up into NearbyTransitPage, and handles updates through a callback,
rather than passing a MutableState object directly.