fix: Fix GL headsign labels on the GLX
Since some atypical patterns from every westbound branch serve the GLX
branches, now that atypical patterns are included to handle properly
displaying those direction labels in predictions, it was causing
predictions on the GLX to always be grouped under a direction label,
even if only typical routes serving the expected headsign were running
at the moment. This takes each grouped static direction, and checks it
against typical and realtime patterns. If there is only a single typical
and/or predicted route going in a direction, then the grouped direction
is instead split up into however many headsigns are typical or predicted
in that direction.