Disable RTL support (#725)
At the moment, the app doesn't support any languages
that are written from right to left, such as Hebrew
or Arabic.

However, AndroidManifest.xml says android:supportsRtl="true",
which is the default value. This causes the app to display
incorrectly in English on phones that are set to work
in a right to left language.

This patch sets rtlandroid:supportsRtl="false",
so that the app won't try to flip itself for
right to left languages unnecessarily.

If full support for such languages is ever added,
set it back to "true".

Co-authored-by: Kayla Brady <31781298+KaylaBrady@users.noreply.github.com>