fix: Make the cache not erase unknown backend fields (#836)
* fix: Write response body directly to cache

Before this, we were writing a re-encoded T that was parsed from the
response body. This resulted in fields which were added to the backend
but didn't exist on the frontend to be erased when cached, but then
when a field was added on the frontend, since the etag was the same
it would continue pulling from a cache which was missing that field
and never update the data.

* fix: Add invalidation key to wipe cache on update

* test: Add tests for caching updates

* fix: Add comment and invalidate shape cache