fix(iOS): Mapbox attribution positioning (#1044)
* fix(iOS): Mapbox attribution positioning

The position wasn't updated when we changed the sheet detent. This
should better handle any future detent ratio changes, and stops using
the deprecated `UIScreen.main.bounds`, which was giving a number that
didn't line up with the actual rendered content height. I tried this on
an emulated iPhone 16 and an iPhone 13 mini to check that it works on
different screen sizes.

* test: Fix tests using explicit hierarchy