chore: Update readme with app store links (#1406)
I realized that the readme never actually said "MBTA Go" anywhere, and figured I'd add the app store links in there while I was at it. I tried to be cute and add the app icon in, but the sizing and built in padding of the resources we have in the repo don't flow well. Maybe we could have Adam make a banner image if we wanted to be over the top.