fix: Retry backend requests (#1458)
I was seeing regular schedule request timeouts, which generally resolve
themselves when you hit the refresh button in the error banner. We don't
need to make users hit refresh and can just retry the requests by
default. This will result in a higher volume of requests to the backend,
especially when there's an issue that causes most requests to fail, but
I think it's still worth it and will lead to riders seeing significantly
fewer unnecessary error banners.