fix: Prevent false positive stale predictions
If trip details was opened, it would continue checking for stale
predictions on a timer indefinitely, even after you go back to a
previous page. This was resulting in the stale banner flickering in and
out on nearby transit, because it was still checking for predictions
which were no longer updating.