fix: Stale predictions flickering and error banner UI (#1471)
* fix: Add error banner to trip details page

It was missing on both platforms. Also update some styling to be more in
line with the rest of the designs (we don't have spec for the banner).

* fix: Prevent false positive stale predictions

If trip details was opened, it would continue checking for stale
predictions on a timer indefinitely, even after you go back to a
previous page. This was resulting in the stale banner flickering in and
out on nearby transit, because it was still checking for predictions
which were no longer updating.